Reorganise fixtures
